backend/requirements.txt
# FastAPI Backend Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
boto3>=1.26.0
botocore>=1.29.0
